requirements.txt
behave==1.2.6
selenium==4.3.0
Appium-Python-Client==4.2.0